requirements.txt
flask>=3.0
pyyaml>=6.0
pytest>=8.0
pytest-benchmark>=4.0
networkx>=3.0
matplotlib>=3.7
pydot>=1.4
z3-solver>=4.12
